refactor: entity -> domain 으로 의미 명확화
