+++ src/main/java/roomescape/reservation/repository/ReservationRepository.java
@@ -2,7 +2,7 @@ package roomescape.reservation.repository;
 
 import java.util.List;
 import java.util.Optional;
-import roomescape.reservation.entity.Reservation;
+import roomescape.reservation.domain.Reservation;
 
 public interface ReservationRepository {
