+++ src/test/java/roomescape/reservation/domain/ReservationTest.java
@@ -1,0 +1,5 @@
+package roomescape.reservation.domain;
+
+public class ReservationTest {
+
+}
