--- src/test/java/roomescape/reservation/entity/ReservationTest.java
@@ -1,9 +1,0 @@
-package roomescape.reservation.entity;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-public class ReservationTest {
-
-
-}
